+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 unreleased
 ----------
+* Add: StringQuery.notInclude (#12)
 * Add: toStringにverboseオプションを追加
 
 3.1.1 / 2017-11-05